+++ src/console_cmds.cpp
@@ -2317,7 +2317,7 @@ DEF_CONSOLE_CMD(ConBankruptCompany)
 	c->bankrupt_value = 42;
 	c->bankrupt_asked = 1 << c->index; // Don't ask the owner
 	c->bankrupt_timeout = 0;
-	c->money = -(UINT64_MAX >> 2);
+	c->money = -static_cast<int64_t>(UINT64_MAX >> 2);
 	IConsolePrint(CC_DEFAULT, "Company marked as bankrupt.");
 
 	return true;
