+++ src/network/core/packet.h
@@ -76,7 +76,7 @@ public:
 	void ReadRawPacketSize();
 	void PrepareToRead();
 
-	bool   CanReadFromPacket (uint bytes_to_read, bool non_fatal = false);
+	bool   CanReadFromPacket(size_t bytes_to_read, bool non_fatal = false);
 	bool   Recv_bool  ();
 	uint8  Recv_uint8 ();
 	uint16 Recv_uint16();
