+++ src/programmable_signals_gui.cpp
@@ -614,7 +614,7 @@ private:
 
 	void RebuildInstructionList()
 	{
-		uint old_len = this->instructions.size();
+		size_t old_len = this->instructions.size();
 		this->instructions.clear();
 		SignalInstruction *insn = program->first_instruction;
 		uint indent = 0;
@@ -677,7 +677,7 @@ private:
 			}
 		} while (insn);
 
-		this->vscroll->SetCount(this->instructions.size());
+		this->vscroll->SetCount(static_cast<int>(this->instructions.size()));
 		if (this->instructions.size() != old_len)
 			selected_instruction = -1;
 		UpdateButtonState();
